+++ core/src/test/java/org/elasticsearch/test/NoopDiscovery.java
@@ -1,0 +1,109 @@
+/*
+ * Licensed to Elasticsearch under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.elasticsearch.test;
+
+import org.elasticsearch.cluster.ClusterChangedEvent;
+import org.elasticsearch.cluster.node.DiscoveryNode;
+import org.elasticsearch.cluster.routing.RoutingService;
+import org.elasticsearch.common.Nullable;
+import org.elasticsearch.common.component.Lifecycle;
+import org.elasticsearch.common.component.LifecycleListener;
+import org.elasticsearch.discovery.Discovery;
+import org.elasticsearch.discovery.DiscoveryStats;
+import org.elasticsearch.discovery.InitialStateDiscoveryListener;
+import org.elasticsearch.node.service.NodeService;
+
+public class NoopDiscovery implements Discovery {
+
+
+    @Override
+    public DiscoveryNode localNode() {
+        return null;
+    }
+
+    @Override
+    public void addListener(InitialStateDiscoveryListener listener) {
+
+    }
+
+    @Override
+    public void removeListener(InitialStateDiscoveryListener listener) {
+
+    }
+
+    @Override
+    public String nodeDescription() {
+        return null;
+    }
+
+    @Override
+    public void setNodeService(@Nullable NodeService nodeService) {
+
+    }
+
+    @Override
+    public void setRoutingService(RoutingService routingService) {
+
+    }
+
+    @Override
+    public void publish(ClusterChangedEvent clusterChangedEvent, AckListener ackListener) {
+
+    }
+
+    @Override
+    public DiscoveryStats stats() {
+        return null;
+    }
+
+    @Override
+    public int getMinimumMasterNodes() {
+        return -1;
+    }
+
+    @Override
+    public Lifecycle.State lifecycleState() {
+        return null;
+    }
+
+    @Override
+    public void addLifecycleListener(LifecycleListener listener) {
+
+    }
+
+    @Override
+    public void removeLifecycleListener(LifecycleListener listener) {
+
+    }
+
+    @Override
+    public Discovery start() {
+        return null;
+    }
+
+    @Override
+    public Discovery stop() {
+        return null;
+    }
+
+    @Override
+    public void close() {
+
+    }
+}
